+++ src/components/LoadingSpinner/LoadingSpinner.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled, { withTheme } from 'styled-components';
-import { fullCircleRotate } from '../../constants/Animations';
+import { fullCircleRotate } from '../../animations';
 import defaultTheme from '../../theme/default';
 
 const colors = {
-  primary: defaultTheme.skyBlue,
-  secondary: defaultTheme.stoneGrey,
-  success: defaultTheme.limeGreen,
-  danger: defaultTheme.flameRed,
-  warning: defaultTheme.apricotOrange,
-  light: defaultTheme.pearlWhite,
-  dark: defaultTheme.ravenBlack,
+  primary: defaultTheme.primary,
+  secondary: defaultTheme.dark,
+  success: defaultTheme.success,
+  danger: defaultTheme.danger,
+  warning: defaultTheme.warning,
+  light: defaultTheme.lighter,
+  dark: defaultTheme.darkest,
 };
 
 const StyledSpinner = styled.span`
@@ -22,19 +22,20 @@ const StyledSpinner = styled.span`
   border: calc(${props => props.size} / 10) solid ${props => props.color};
   border-left-color: ${props => props.gapColor};
   border-radius: 100%;
-  animation: ${fullCircleRotate} ${props => props.speed} linear infinite;
+  animation: ${fullCircleRotate()} ${props => props.speed} linear infinite;
 `;
 
-const LoadingSpinner = ({ size, speed, appearance, ...props }) => {
+const LoadingSpinner = ({ size, speed, appearance, ...otherProps }) => {
   const defaultColor = colors[appearance];
   const defaultGapColor = `${defaultColor}25`;
 
   return (
     <StyledSpinner
+      {...otherProps}
       size={size}
-      color={props.color || defaultColor}
+      color={otherProps.color || defaultColor}
       speed={speed}
-      gapColor={props.color ? (props.gapColor || 'transparent') : defaultGapColor}
+      gapColor={otherProps.color ? (otherProps.gapColor || 'transparent') : defaultGapColor}
     />
   );
 };