+++ src/helpers/styled.js
@@ -1,0 +1,9 @@
+export const paddingHorizontal = value => `
+  padding-left: ${value};
+  padding-right: ${value};
+`;
+
+export const paddingVertical = value => `
+  padding-top: ${value};
+  padding-bottom: ${value};
+`;