+++ src/theme/colors.js
@@ -1,0 +1,17 @@
+// TODO: Use more abstract names like colorPrimary, colorSecondary, colorWarning, colorError (?)
+const colors = {
+  darkest: '#313233', // ravenBlack
+  darker: '#595959',
+  dark: '#9b9b9b', // stoneGrey
+  light: '#e6e6e6',
+  lighter: '#f7f8fa', // pearlWhite
+  lightest: '#ffffff', // white
+  disabled: '#c9ccd1',
+  primary: '#5f9dc7', // skyBlue
+  primaryActive: '#337aa9',
+  danger: '#e25454', // flameRed
+  success: '#aecc76', // limeGreen
+  warning: '#f4a671', // apricotOrange
+};
+
+export default colors;