+++ webpack.production.config.js
@@ -6,7 +6,7 @@ const { LicenseWebpackPlugin } = require('license-webpack-plugin');
 
 
 module.exports = {
-  entry: path.resolve(__dirname, './src/components/index.js'),
+  entry: path.resolve(__dirname, './src/index.js'),
   output: {
     path: path.resolve(__dirname, './dist'),
     filename: 'index.js',